// web-frontend/norbitz/src/apis/carnivore-cruise-lines/model/cruiseItem.ts
/**
 * Carnivore Cruise Lines API
 * This is an API to access the information on Carnivore Cruise Lines
 *
 * OpenAPI spec version: 1.3.4
 * Contact: jcw931@msstate.edu
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CruiseItem {
    available: boolean;
    cost: number;
    cruiseLinerID: string;
    departureDate: string;
    description: string;
    duration: number;
    fromLocation: string;
    itemID: string;
    name: string;
    returnDate: string;
    roomCapacity: number;
    roomID: number;
}